requirements.txt
xarray==2022.11.0
cfgrib==0.9.10.3
geopandas==0.11.1
numpy==1.24.0
scipy==1.9.3
python_dotenv